// src/app/models/ims/adjustment-report.ts
export class AdjustmentReport {
    barcode: string;
    productName: string;
    area: string;
    adjustmentNumber: string;
    adjustmentDate: string;
    adjustmentType: string;
    adjustmentMotiveGroup: string;
    motive: string;
    inventory: number;
    count: number;
    totalIn: number;
    totalOut: number;
    cost: number;
    price: number;
    operator: string;
    responsible: string;
    category: string;
}